Remove API transport file
